seperated left side bar from dashboard layout
